requirements.txt
requests==2.25.0
pyyaml==5.3.1
